move ZkTrie to trie package
