+++ trie/zk_trie_node.go
@@ -1,4 +1,4 @@
-package zktrie
+package trie
 
 import (
 	"encoding/binary"
